refactor: auth, config, controller 내부 패키지 수정
