+++ src/main/java/cart/auth/resolver/BasicAuthenticationPrincipal.java
@@ -1,4 +1,4 @@
-package cart.controller.auth.resolver;
+package cart.auth.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
