+++ src/test/java/cart/JwpCartApplicationTests.java
@@ -1,7 +1,7 @@
 package cart;
 
-import cart.controller.dto.ItemRequest;
-import cart.controller.dto.ItemResponse;
+import cart.controller.dto.request.ItemRequest;
+import cart.controller.dto.response.ItemResponse;
 import cart.domain.ImageUrl;
 import cart.domain.Item;
 import cart.domain.Name;
